Fix RSS icon bug in the footer
This commit fixes a bug where the RSS icon was not showing in the
footer. To implement it I simply added a new RSS icon to the defs.svg
file. I got the svg from a site called simpleicons.org which is
completely open source and free to use. I also edited the styling a bit
to make the icon smaller and move it a bit to the left, so that it is a
similar height to the text and looks more centered.