Add Dribbble and Behance to social network options
This commit adds the design social networks dribbble and behance to the
options of social networks in the customizer. To do this I added the
icons into the defs.svg file. I got the icons from a site called
simpleicons.org which is open source and completely free to use. Then i
added then I added the fields to the customizer in the customizer.php
file. Then I added the string translations for the customizer label.
Finally I added the social network names to the social_names array in
profile.php